requirements.txt
torch>=2.0.0
torchvision
lightning>=2.0.0
git+https://github.com/openai/CLIP.git
lovely-tensors
hydra-core==1.3.2
hydra-colorlog==1.2.0
rootutils
wandb
omegaconf
rich
loguru